requirements.txt
tensorflow>=2.10.0
matplotlib>=3.5.0
scikit-learn>=1.1.0
seaborn>=0.11.0
numpy>=1.21.0
# Core ML and Data Processing
opencv-python>=4.5.0
Pillow>=8.3.0

# Streamlit and UI
streamlit>=1.28.0
streamlit-drawable-canvas>=0.9.0


